+++ services/API_ClayTrack/Models/DetailSale.cs
@@ -10,7 +10,7 @@ namespace API_ClayTrack.Models
         public int idDetailSale { get; set; }
 
         [Required]
-        public float quantity { get; set; }
+        public int quantity { get; set; }
 
         [Required]
         public float price { get; set; }
